% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscutils.R
\name{are.attributes}
\alias{are.attributes}
\title{check that all non-missing attributes are character and are valid inexs into the data}
\usage{
are.attributes(attrs, data)
}
\description{
check that all non-missing attributes are character and are valid inexs into the data
}
