% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscutils.R
\name{are.operator.value}
\alias{are.operator.value}
\title{check that all attributes are greater than zero}
\usage{
are.operator.value(attrs, opr = `>=`, val = 0)
}
\description{
check that all attributes are greater than zero
}
\keyword{internal}
