% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extradplyr.R
\name{attribute.class}
\alias{attribute.class}
\title{get the class for each column in data, as a named character vector}
\usage{
attribute.class(data, .checkAssertions = T)
}
\arguments{
\item{data}{a data frame or tibble type object}
}
\value{
a character vector of the class type of each column
}
\description{
get the class for each column in data, as a named character vector
}
\examples{
chr.ex <- sample(c('a','b','c'), 100, TRUE)
int.ex <- sample(c(1,2,3), 100, TRUE)
dttm.ex <- sample(as.Date.character(paste('2020-01-', 1:30, sep = '')), 100, TRUE)
sample.data <- data.frame('chr' = chr.ex, 'int' = int.ex, 'dttm' = dttm.ex)

# get data classes for all columns
attribute.class(sample.data)
}
