% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extradplyr.R
\name{attribute_class}
\alias{attribute_class}
\title{Get class type for all columns in a data frame or tibble}
\usage{
attribute_class(data, .checkAssertions = TRUE)
}
\arguments{
\item{data}{a data frame or tibble type object}

\item{.checkAssertions}{private variable used to check function assertions (if first in chain)}
}
\value{
a character vector of the class type of each column
}
\description{
get the class for each column in data, as a named character vector
}
\examples{
chr.ex <- sample(c('a','b','c'), 100, TRUE)
int.ex <- sample(c(1,2,3), 100, TRUE)
dttm.ex <- sample(as.Date.character(paste('2020-01-', 1:30, sep='')), 100, TRUE)
sample.data <- data.frame('chr'=chr.ex, 'int'=int.ex, 'dttm'=dttm.ex)

# get data classes for all columns
attribute_class(sample.data)
}
