% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extradplyr.R
\name{attribute_stats}
\alias{attribute_stats}
\title{Attribute Stats}
\usage{
attribute_stats(data, stats_attr = NULL, .checkAssertions = TRUE)
}
\arguments{
\item{data}{a data frame or tibble type object}

\item{stats_attr}{a column label as a string to get stats for}

\item{.checkAssertions}{private variable used to check function assertions (if first in chain)}
}
\value{
a tibble with the summary stats for the specified column/grouping
}
\description{
get the summary statistics for an attribute (grouped or otherwise)
}
\examples{
attribute_stats( iris )
attribute_stats( iris, 'Petal.Length')
}
