% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extradplyr.R
\name{attributes_summarise}
\alias{attributes_summarise}
\title{Summarise Attributes}
\usage{
attributes_summarise(data, group_attr = NULL, .checkAssertions = TRUE)
}
\arguments{
\item{data}{a data frame or tibble type object}

\item{group_attr}{a column label as a string to group by before summarizing (optional)}

\item{.checkAssertions}{private variable used to check function assertions (if first in chain)}
}
\value{
a list with stats for each attribute (grouped by group_attr if present)
}
\description{
generate a summary of each attribute in the dataset grouped by a target attribute
}
\examples{
attributes_summarise( iris )
attributes_summarise( iris, 'Species' )
}
