% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extradplyr.R
\name{attributes_summarize}
\alias{attributes_summarize}
\title{Summarize Attributes}
\usage{
attributes_summarize(data, group_attr = NULL, .checkAssertions = TRUE)
}
\arguments{
\item{data}{a data frame or tibble type object}

\item{group_attr}{a column label as a string to group by before summarizing (optional)}

\item{.checkAssertions}{private variable used to check function assertions (if first in chain)}
}
\value{
a list with stats for each attribute (grouped by group_attr if present)
}
\description{
generate a summary of each attribute in the dataset grouped by a target attribute
(helper function for attribute summary)
}
\examples{
attributes_summarize(iris)
attributes_summarize(iris, 'Species')
}
\keyword{internal}
