% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{classification_metrics}
\alias{classification_metrics}
\title{Classification Metrics}
\usage{
classification_metrics(mod, y, y.hat, pi_0 = 0.5)
}
\arguments{
\item{mod}{a general linear model of family binomial}

\item{y}{actual classification values ( only used if mod not specified )}

\item{y.hat}{model predicitions as percentage ( only used if mod not specified )}

\item{pi_0}{the threshold of determination used in classifaction ( optional, default = 0.5 )}
}
\value{
A list of classification performance metrics derived from the confusion matrix
\itemize{
 \item{y.hat}{      : the model classification predictions}
 \item{confmatrix}{ : the confusion matrix}
 \item{tpos}{       : count of true positives}
 \item{tneg}{       : count of true negatives}
 \item{fneg}{       : count of false negatives}
 \item{fpos}{       : count of false positives}
 \item{acc}{        : accuracy score}
 \item{recall}{     : recall score}
 \item{precision}{  : precision score}
 \item{tpr}{        : true positive rate (same as recall)}
 \item{fpr}{        : false positive rate}
 \item{fdcr}{       : false discovery rate}
 \item{fnr}{        : false negative rate}
 \item{fomr}{       : false omission rate}
 \item{tnr}{        : true negative rate}
 \item{ppv}{        : positive predicted value (same as precision)}
 \item{npv}{        : negative predicted value}
 \item{F1}{         : F1 scoring metric}
}
}
\description{
extract all performance metrics for a binomial classification model
}
\examples{
X <- rnorm(100)
Y <- ifelse(X < 0.5, 0, 1 - sample(0:1, prob = c(0.9, .1)))
f <- glm(Y ~ X, family='binomial')
classification_metrics(f)
}
