% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscutils.R
\name{combine_lists}
\alias{combine_lists}
\title{Combine Lists}
\usage{
combine_lists(l1, ..., merge = FALSE)
}
\arguments{
\item{l1}{a starting list}

\item{...}{a list or lists to combine}

\item{merge}{should matching list elements be merged together or overridden by latest list (default=FALSE)}
}
\value{
a single list with all the elements appearing in the input lists
}
\description{
combine multiple successive lists together
}
\examples{
L1 <- list(a = 1:10, b=11:20)
L2 <- list(c = 21:30, d=31:40)
L3 <- list(e = 41:50, a=51:60)
L4 <- combine_lists(L1, L2)
L5 <- combine_lists(L1, L2, L3)
L6 <- combine_lists(L3, L4, L5, L1)
L7 <- combine_lists(L1, L2, L3, merge=TRUE)
}
