% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{data_split}
\alias{data_split}
\title{Split data into specified proportions}
\usage{
data_split(data, prop)
}
\arguments{
\item{data}{a data.frame type object}

\item{prop}{the proportion of the data in the secondary (smaller) subset}
}
\value{
a list with the primary and secondary subsets in specified proportion
}
\description{
helper function for splitting out data sets by proportion
}
\examples{
data_split(cars, 0.5)
}
