% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extradplyr.R
\name{diversity}
\alias{diversity}
\title{Entropy Based Measures of Data Diversity}
\usage{
diversity(data, ...)
}
\arguments{
\item{data}{a data frame or tibble type object}

\item{...}{arguments passed to 'entropy' function}
}
\value{
a named vector containing the percentage diversity,
the true entropy of the dataset,
and the potential entropy of the dataset
( diversity, entropy, potential.entropy )
}
\description{
Calculate the diversity of a dataset by determining the degree to which
differentiable groupings in the columns vary from maximum entropy for the number of grouping.
If the groupings are uniformly distributed, the entropy is maximized.  This functions calculates
the ratio of the real distribution of groupings to the maximum entropy distribution to determine
the overall diversity represented in the grouped data.
}
\examples{
diversity(cars)

t <- data.frame(a = rep(0, 100))
diversity(t)
}
