% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscutils.R
\name{.assertions}
\alias{.assertions}
\title{Assertions used to change data entries and attributes}
\usage{
.assertions(data, attributes = NULL, requires_attributes = F, isRun = T)
}
\description{
Assertions used to change data entries and attributes
}
