% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscutils.R
\name{get_attribute}
\alias{get_attribute}
\title{Get Attribute From List or Data Frame}
\usage{
get_attribute(x, attribute, default = NULL)
}
\arguments{
\item{x}{a list-like object (e.g. data.frame)}

\item{attribute}{the name or index of an element in x}

\item{default}{value to return if attribute is missed or error (NULL)}
}
\value{
the value of attribute from x or default
}
\description{
select a named element from a list or data.frame with default for missed values
}
\examples{
\dontrun{
y <- list('a' = 1)
x <- data.frame('b' = 2)
get_attribute(y, 'a', 2) # 1
get_attribute(y, 'b', 1) # 1
get_attribute(x, 'a', 2) # 2
get_attribute(x, 'b', 1) # 2
}
}
