% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscutils.R
\name{get_indices}
\alias{get_indices}
\title{Get Data Indices}
\usage{
get_indices(data)
}
\arguments{
\item{data}{any data set with positional elements}
}
\value{
either a numeric vector of positional indices or a character vector of named entries
}
\description{
returns a vector with the indices into the dataset,
either as the named entries or the ordered positions of entries
}
\examples{
get_indices(100:110)
get_indices(c('a'=10,'b'=20))
get_indices(iris)
}
