% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraggplot.R
\name{gghistogram}
\alias{gghistogram}
\title{Generate quick informative summary histograms}
\usage{
gghistogram(
  data,
  summary_attr = NULL,
  group_attr = NULL,
  ...,
  fill_color = "blue",
  line_color = "red"
)
}
\arguments{
\item{data}{the dataset, a data.frame or tibble object}

\item{summary_attr}{the attribute to show (x - variable)}

\item{group_attr}{the grouping attribute (fill - variable)}

\item{...}{arguments passed to geom_histogram, (e.g. bins, binwidth)}

\item{fill_color}{the default fill}

\item{line_color}{the default line color}
}
\value{
a ggplot object to display
}
\description{
Special format for histograms showing the summary statistics on the x axis, and the mean as a line
For numerical data, generates vertical mean lines for each group.  For categorical data, generates a horizontal line
representing the mean count.
}
