% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscutils.R
\name{is.attribute}
\alias{is.attribute}
\title{check that non-missing attribute argument is character and is a valid index into the data}
\usage{
is.attribute(attr, nms)
}
\description{
check that non-missing attribute argument is character and is a valid index into the data
}
\keyword{internal}
