% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscutils.R
\name{is.bidirected}
\alias{is.bidirected}
\title{Is Bidirected}
\usage{
is.bidirected(x)
}
\arguments{
\item{x}{an object to test}
}
\value{
TRUE if x is Bidirected
}
\description{
bidirected types exist on a continuum with positive and negative values
with no preferred direction (as opposed to temporal types where time moves forward, for example)
}
\examples{
is.bidirected(c(1,2,3)) # TRUE
is.bidirected(c('1','2','3')) # FALSE
}
