% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscutils.R
\name{is.categorical}
\alias{is.categorical}
\title{check if x is categorical type}
\usage{
is.categorical(x)
}
\arguments{
\item{x}{an object to test}
}
\value{
TRUE if x is Categorical
}
\description{
categorical types do not have orderings or hierarchies
and cannot be compared within set (as opposed to comparable types)
}
\examples{
is.categorical(c('1','2','3')) # TRUE
is.categorical(Date(1)) # FALSE
}
