% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscutils.R
\name{is.comparable}
\alias{is.comparable}
\title{Is Comparable}
\usage{
is.comparable(x)
}
\arguments{
\item{x}{an object to test}
}
\value{
TRUE if x is Comparable
}
\description{
comparable types have orderings that can be used for sorting
and comparison within set (as opposed to categorical types)
}
\examples{
is.comparable(c(1,2,3)) # TRUE
is.comparable(c('1','2','3')) # FALSE
}
