% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscutils.R
\name{is.falsey}
\alias{is.falsey}
\title{Check if x is false-like}
\usage{
is.falsey(x)
}
\arguments{
\item{x}{an object to test}
}
\value{
TRUE is x is falsey
}
\description{
falsey checks if x should be interpreted like FALSE,
it returns true if x is any of FALSE, NULL, NA, 0, 'FALSE', length zero, missing, or error
}
\examples{
is.falsey(FALSE)
is.falsey(TRUE)
is.falsey(NULL)
is.falsey(NA)
is.falsey(0)
is.falsey(1)
is.falsey(-1)
is.falsey('F')
is.falsey('FALSE')
is.falsey('F ')
is.falsey(logical(0))
is.falsey(c())
}
