% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscutils.R
\name{is.temporal}
\alias{is.temporal}
\title{Is Temporal}
\usage{
is.temporal(x)
}
\arguments{
\item{x}{an object to test}
}
\value{
TRUE if x is Temporal
}
\description{
temporal types are references to time and date
}
\examples{
is.temporal(Date(1)) # TRUE
is.temporal(c(1,2,3)) # FALSE
}
