% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscutils.R
\name{ops_util}
\alias{ops_util}
\title{utility function for formatting sets of operations}
\usage{
ops_util(x, operator)
}
\description{
utility function for formatting sets of operations
}
\keyword{internal}
