% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extradplyr.R
\name{uniqueCounts}
\alias{uniqueCounts}
\title{Unique Group Counts}
\usage{
uniqueCounts(data, include.percent = T, .checkAssertions = T)
}
\arguments{
\item{data}{a data frame or tibble type object}

\item{include.percent}{whether a column should be added to show group percentages}
}
\value{
a grouped data frame of unique rows with column count ( and percent )
}
\description{
Unique Group Counts
}
\examples{
uniqueCounts( cars )
uniqueCounts(iris, T)
}
