% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscutils.R
\name{update_value}
\alias{update_value}
\title{Update Value}
\usage{
update_value(value, update, miss = is.null)
}
\arguments{
\item{value}{the current value of the variable being updated}

\item{update}{the updated value to check}

\item{miss}{the condition function for update ( default = is.null )}
}
\value{
update if miss condition not met, otherwise value
}
\description{
set value to update only if update passes miss condition
}
\examples{
y<- 100
y<-update_value(y, 200)
y<-update_value(y, NULL)
y<-update_value(y, 'a', function(x) x=='a')
}
